Infinite scroll working for posts and members inside group page.
- Allowed multiple infinite scrolls at the same time
- Need to use the correct view when loading the users
- Most likely need to add event listeners to newly loaded users (e.g. buttons to remove/add for the owner)

+++ public/js/infinite_scrolling.js
@@ -30,14 +30,18 @@ async function createFetcher(container, testIntersectionElement, url, firstActio
   observer.observe(testIntersectionElement);
 }
 
-export async function infiniteScroll(container, testIntersectionElement, url, firstAction, action, data = {}) {
-  container.dataset.page = 0;
-  destroyFetcher();
+export async function infiniteScroll(container, testIntersectionElement, url, firstAction, action, resetPage = true, destroyPreviousFetcher = true, data = {}) {
+  if (resetPage) {
+    container.dataset.page = 0;
+  }
+  if (destroyPreviousFetcher) {
+    destroyFetcher();
+  }
   createFetcher(container, testIntersectionElement, url, firstAction, action, data);
 }
 
 export function destroyFetcher() {
   if (observer) {
-    observer.disconnect(); // Disconnect the observer if it exists
+    observer.disconnect();
   }
 }
